Split compose discover form steps
Move the compose discover stepped form and related tests into
step-scoped files so the form stays maintainable as it grows.

compose_discover_form/ now contains:
- compose_discover_form.tsx — orchestrator (step registry + form component)
- alert_condition_step.tsx
- recovery_condition_step.tsx
- details_and_artifacts_step.tsx
- notifications_step.tsx
- index.ts — barrel re-export
