+++ x-pack/platform/packages/shared/response-ops/alerting-v2-rule-form/flyout/compose_discover/compose_discover_form/alert_condition_step.tsx
@@ -1,0 +1,242 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import React, { useCallback, useEffect, useMemo, useRef } from 'react';
+import { useFormContext } from 'react-hook-form';
+import { Parser, isColumn } from '@elastic/esql';
+import { useQuery } from '@kbn/react-query';
+import { getEsqlColumns } from '@kbn/esql-utils';
+import {
+  EuiButton,
+  EuiCallOut,
+  EuiComboBox,
+  EuiFormRow,
+  EuiPanel,
+  EuiSelect,
+  EuiSpacer,
+  EuiSwitch,
+  EuiText,
+  EuiTitle,
+} from '@elastic/eui';
+import type { ComposeDiscoverAction, ComposeDiscoverState } from '../types';
+import type { ComposeFormValues } from '../compose_form_types';
+import { QuerySummary } from '../query_summary';
+import type { RuleFormServices } from '../../../form/contexts/rule_form_context';
+import { useDataFields } from '../../../form/hooks/use_data_fields';
+import { ScheduleField } from '../../../form/fields/schedule_field';
+import { LookbackWindowField } from '../../../form/fields/lookback_window_field';
+
+interface AlertConditionStepProps {
+  state: ComposeDiscoverState;
+  dispatch: React.Dispatch<ComposeDiscoverAction>;
+  services: RuleFormServices;
+}
+
+export function AlertConditionStep({ state, dispatch, services }: AlertConditionStepProps) {
+  const { setValue, watch } = useFormContext<ComposeFormValues>();
+  const isAlert = watch('kind') === 'alert';
+  const timeField = watch('timeField') ?? '@timestamp';
+  const grouping = watch('grouping');
+  const groupFields = grouping?.fields ?? [];
+  const query = watch('query');
+
+  const baseQuery = query.format === 'composed' ? query.base : '';
+  const alertBlock = query.format === 'composed' ? query.blocks.breach : '';
+  const fullQuery = query.format === 'standalone' ? query.breach : '';
+
+  // Use the base query for field lookup when tracking is on; fall back to full breach query.
+  const committedQuery = useMemo(() => {
+    const q = isAlert ? baseQuery : fullQuery;
+    return /^\s*FROM\s+[a-zA-Z0-9_.*-]/i.test(q) && state.queryCommitted ? q : '';
+  }, [isAlert, baseQuery, fullQuery, state.queryCommitted]);
+
+  const { data: fieldMap } = useDataFields({
+    query: committedQuery,
+    http: services.http,
+    dataViews: services.dataViews,
+  });
+  const timeFieldOptions = useMemo(() => {
+    const dateFields = Object.values(fieldMap ?? {})
+      .filter((f) => f.type === 'date')
+      .map((f) => f.name)
+      .sort();
+    if (!dateFields.includes('@timestamp')) dateFields.unshift('@timestamp');
+    return dateFields.map((name) => ({ value: name, text: name }));
+  }, [fieldMap]);
+
+  // Output columns of the full pipeline -> options for the group fields selector.
+  // Uses | LIMIT 0 so no data is transferred -- only the output schema is returned.
+  // Works in edit mode (query seeded on mount) without requiring the sandbox to be opened.
+  const { data: outputColumns = [] } = useQuery({
+    queryKey: ['composeDiscoverOutputColumns', committedQuery],
+    queryFn: async () => {
+      const cols = await getEsqlColumns({
+        esqlQuery: committedQuery,
+        search: services.data.search.search,
+      });
+      return cols.map((c) => c.name);
+    },
+    enabled: Boolean(committedQuery),
+    refetchOnWindowFocus: false,
+    keepPreviousData: true,
+  });
+
+  // Auto-populate group fields from the STATS BY clause when a query is first committed
+  // and the user hasn't already set any group fields.
+  const autoPopulatedForRef = useRef<string | null>(null);
+  useEffect(() => {
+    if (!state.queryCommitted || groupFields.length > 0 || !committedQuery) return;
+    if (autoPopulatedForRef.current === committedQuery) return;
+    autoPopulatedForRef.current = committedQuery;
+    try {
+      const { root } = Parser.parse(committedQuery);
+      const statsCmd = [...root.commands].reverse().find((c) => c.name === 'stats');
+      // ESQLAstItem is a wide union -- use a local type alias to access the 'by' option
+      // safely rather than an inline interface (which triggers lint in function scope).
+      interface CmdOption {
+        type: string;
+        name: string;
+        args?: unknown[];
+      }
+      const byOption = (statsCmd?.args as CmdOption[] | undefined)?.find(
+        (a) => a.type === 'option' && a.name === 'by'
+      );
+      const byFields = (byOption?.args ?? []).filter(isColumn).map((a) => a.name);
+      if (byFields.length > 0) setValue('grouping', { fields: byFields });
+    } catch {
+      // Non-parseable query -- skip auto-populate
+    }
+  }, [state.queryCommitted, committedQuery, groupFields.length, setValue]);
+
+  const handleTrackingToggle = useCallback(() => {
+    setValue('kind', isAlert ? 'signal' : 'alert');
+  }, [isAlert, setValue]);
+
+  // Callout when the heuristic split couldn't find a clear split point.
+  // Only relevant after Apply (when the committed query is in composed format).
+  const splitFailed =
+    isAlert && state.queryCommitted && query.format === 'composed' && !query.base.trim();
+
+  return (
+    <>
+      <EuiTitle size="xs">
+        <h3>ES|QL query</h3>
+      </EuiTitle>
+      <EuiSpacer size="s" />
+
+      {!state.queryCommitted ? (
+        <>
+          <EuiPanel color="subdued" paddingSize="m">
+            <EuiText size="s" color="subdued">
+              No query defined yet
+            </EuiText>
+          </EuiPanel>
+          <EuiSpacer size="s" />
+          <EuiButton
+            iconType="editorCodeBlock"
+            isDisabled={state.childOpen}
+            onClick={() => dispatch({ type: 'OPEN_CHILD_FOR_STEP', step: state.step })}
+            data-test-subj="composeDiscoverOpenEditor"
+          >
+            Open query editor
+          </EuiButton>
+        </>
+      ) : !isAlert ? (
+        <>
+          <QuerySummary query={fullQuery} label="query" />
+          <EuiSpacer size="s" />
+          <EuiButton
+            size="s"
+            iconType="editorCodeBlock"
+            isDisabled={state.childOpen}
+            onClick={() => dispatch({ type: 'OPEN_CHILD_FOR_STEP', step: state.step })}
+            data-test-subj="composeDiscoverEditQuery"
+          >
+            Edit query
+          </EuiButton>
+        </>
+      ) : (
+        <>
+          {splitFailed && (
+            <>
+              <EuiCallOut
+                announceOnMount={false}
+                size="s"
+                color="primary"
+                iconType="iInCircle"
+                title="Couldn't automatically separate base query from alert condition. Adjust the split in the query editor."
+              />
+              <EuiSpacer size="s" />
+            </>
+          )}
+          <EuiText size="xs" color="subdued">
+            <strong>Base query</strong>
+          </EuiText>
+          <EuiSpacer size="xs" />
+          <QuerySummary query={baseQuery} label="base query" />
+          <EuiSpacer size="m" />
+          <EuiText size="xs" color="subdued">
+            <strong>Alert condition</strong>
+          </EuiText>
+          <EuiSpacer size="xs" />
+          <QuerySummary query={alertBlock} label="alert condition" />
+          <EuiSpacer size="s" />
+          <EuiButton
+            size="s"
+            iconType="editorCodeBlock"
+            isDisabled={state.childOpen}
+            onClick={() => dispatch({ type: 'OPEN_CHILD_FOR_STEP', step: state.step })}
+            data-test-subj="composeDiscoverEditQueries"
+          >
+            Edit queries
+          </EuiButton>
+        </>
+      )}
+
+      <EuiSpacer size="m" />
+      <EuiFormRow label="Time field" fullWidth>
+        <EuiSelect
+          fullWidth
+          options={timeFieldOptions}
+          value={timeField}
+          onChange={(e) => setValue('timeField', e.target.value)}
+          disabled={state.childOpen}
+          data-test-subj="composeDiscoverTimeField"
+        />
+      </EuiFormRow>
+      <EuiSpacer size="m" />
+      <EuiFormRow label="Group fields" fullWidth>
+        <EuiComboBox
+          fullWidth
+          options={outputColumns.map((name) => ({ label: name }))}
+          selectedOptions={groupFields.map((f) => ({ label: f }))}
+          onChange={(opts) =>
+            setValue('grouping', opts.length ? { fields: opts.map((o) => o.label) } : undefined)
+          }
+          onCreateOption={(val) => setValue('grouping', { fields: [...groupFields, val] })}
+          placeholder="Add group fields"
+          data-test-subj="composeDiscoverGroupFields"
+        />
+      </EuiFormRow>
+
+      <EuiSpacer size="m" />
+      <EuiSwitch
+        label="Track active and recovered state over time"
+        checked={isAlert}
+        onChange={handleTrackingToggle}
+        disabled={!state.queryCommitted}
+        data-test-subj="composeDiscoverTrackingToggle"
+      />
+
+      {/* Schedule and lookback -- connected to RHF via useFormContext() internally */}
+      <EuiSpacer size="m" />
+      <ScheduleField />
+      <EuiSpacer size="m" />
+      <LookbackWindowField />
+    </>
+  );
+}
