+++ x-pack/platform/packages/shared/response-ops/alerting-v2-rule-form/flyout/compose_discover/compose_discover_form/details_and_artifacts_step.tsx
@@ -1,0 +1,48 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import React from 'react';
+import {
+  EuiFieldText,
+  EuiFormRow,
+  EuiHorizontalRule,
+  EuiSpacer,
+  EuiText,
+  EuiTitle,
+} from '@elastic/eui';
+import { RuleDetailsFieldGroup } from '../../../form';
+
+export function DetailsAndArtifactsStep() {
+  return (
+    <>
+      {/* Name, description, tags -- connected to RHF via useFormContext() internally */}
+      <RuleDetailsFieldGroup />
+
+      <EuiHorizontalRule margin="m" />
+
+      <EuiTitle size="xs">
+        <h3>Artifacts</h3>
+      </EuiTitle>
+      <EuiSpacer size="s" />
+
+      {/* TODO (#268770): wire runbook URL and dashboard link to FormValues.artifacts */}
+      <EuiFormRow label="Runbook URL" fullWidth labelAppend={<EuiText size="xs">Optional</EuiText>}>
+        <EuiFieldText fullWidth placeholder="https://..." disabled />
+      </EuiFormRow>
+
+      <EuiSpacer size="m" />
+
+      <EuiFormRow
+        label="Dashboard link"
+        fullWidth
+        labelAppend={<EuiText size="xs">Optional</EuiText>}
+      >
+        <EuiFieldText fullWidth placeholder="https://..." disabled />
+      </EuiFormRow>
+    </>
+  );
+}
