+++ x-pack/platform/packages/shared/response-ops/alerting-v2-rule-form/flyout/compose_discover/compose_discover_form/notifications_step.tsx
@@ -1,0 +1,27 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import React from 'react';
+import { EuiCallOut } from '@elastic/eui';
+
+// TODO (#268770): Notifications step -- wire workflow selector and notification policy fields
+// to FormValues once the action policy API integration is in place.
+export function NotificationsStep() {
+  return (
+    <EuiCallOut
+      title="Notifications configuration coming soon"
+      iconType="clock"
+      color="primary"
+      size="s"
+    >
+      <p>
+        Notification policies will be configurable here. Rules are created without notifications
+        until this step is wired.
+      </p>
+    </EuiCallOut>
+  );
+}
